pyproject.toml
[project]
name = "xai-emotion-recognition"
version = "0.1.0"
description = "Explainable AI for Facial Emotion Recognition — Grounded Textual Explanations via Multimodal Feature Fusion"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

authors = [
    {name = "Abhi"},
]

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.40.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.30.0",

    # Vision & Face Detection
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",

    # Attention Maps
    "pytorch-grad-cam>=1.5.0",

    # Data & Scientific
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",

    # Config & CLI
    "pyyaml>=6.0",
    "tqdm>=4.65.0",

    # Dataset
    "kaggle>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "ipykernel>=6.0",
    "jupyter>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
